requirements.txt
# HB-BIS - Handwriting-Based Biometric Identification System
# Educational Prototype Dependencies

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Deep learning (PyTorch)
torch>=1.9.0
torchvision>=0.10.0

# Machine learning
scikit-learn>=0.24.0

# Visualization (for preprocessing demonstration)
matplotlib>=3.3.0

# Note: Tkinter comes built-in with Python, no installation needed
